refactor: split the surface state types out of the contract
The contract file sat exactly at its line cap, so the two-line drawing
selection member could not land. Its read side moves to its own module and
the contract now fits the global cap. The drawing position lookup gets a
small module of its own rather than growing either caller.
